backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# AI & Agents
anthropic
pydantic>=2.0
pydantic-ai

# Database
supabase
psycopg2-binary

# Code Parsing & Analysis
beautifulsoup4
lxml
esprima
astroid

# Task Queue
celery
redis

# Testing
pytest
pytest-asyncio
httpx

# Utilities
pyyaml
jinja2
aiofiles
